Clean up src\wasm-interpreter.h(307): warning C4244: 'argument': conversion from 'int64_t' to 'unsigned int', possible loss of data

+++ src/wasm-interpreter.h
@@ -304,7 +304,7 @@ private:
 
         Name target = curr->default_;
         if (index >= 0 && (size_t)index < curr->targets.size()) {
-          target = curr->targets[index];
+          target = curr->targets[(size_t)index];
         }
         flow.breakTo = target;
         return flow;
